backend/requirements.txt
# Core Django
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1
argon2-cffi==23.1.0
python-decouple==3.8

# Async & Task Queue
celery==5.3.6
redis==5.0.2
django-redis==5.4.0

# Payment Integrations
stripe==8.7.0
rave-python==1.4.1
paystackapi==2.1.0

# Communication
twilio==9.0.0
africastalking==1.2.8
sendgrid==6.11.0

# File Handling
Pillow==10.2.0
python-magic==0.4.27
boto3==1.34.51

# API Documentation
drf-spectacular==0.27.1

# Development
django-extensions==3.2.3
django-debug-toolbar==4.3.0
ipython==8.22.1

# Testing
pytest==8.0.2
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==23.2.1

# Utilities
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.6.3
pandas==2.2.1
openpyxl==3.1.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.40.6
